Simplify the Disable logic and introduce %SAVED. Readability, formatting.
Instead of modules calling MkSave() which does not emit anything (and can
be misleading when placed in if statements), introduce %SAVED where modules
can register a list of make variables to export and handle it generically
in test() and disable().

Introduce MkDisableNotFound() and MkDisableFailed() for modules. This
allows disable routines to optimize against unnecessary redefinitions
based on context.



git-svn-id: file:///home/dev/bsdbuild/trunk@1246 086a213a-9b02-0410-a1af-a7f5267bcc5c